Normalize current to mAh; Only report remaining energy if value is valid; Temperature to tenths of C to C

+++ lib/model/android_battery_info.dart
@@ -12,7 +12,7 @@ class AndroidBatteryInfo {
   int? batteryCapacity;
   int? remainingEnergy = -1;
   int? scale;
-  int? temperature = -1;
+  double? temperature = -1;
   int? voltage = -1;
   bool? present = true;
   ChargingStatus? chargingStatus;
